+++ transektcount/src/main/java/com/wmstein/transektcount/database/Meta.kt
@@ -2,17 +2,23 @@ package com.wmstein.transektcount.database
 
 /***********************************
  * Created by wmstein on 31.03.2016.
- * Converted to Kotlin on 2023-06-26
+ * Converted to Kotlin on 2024-03-09
  */
 class Meta {
     @JvmField
     var id = 0
     @JvmField
+    var temps = 0
+    @JvmField
     var tempe = 0
     @JvmField
-    var wind = 0
+    var winds = 0
+    @JvmField
+    var winde = 0
     @JvmField
     var clouds = 0
+    @JvmField
+    var cloude = 0
     @JvmField
     var date: String? = null
     @JvmField